+++ inc/staffSvg/scheduler.ts
@@ -72,6 +72,7 @@ export default class Scheduler {
 			const nextItem = tickTable[i + 1];
 
 			item.endTick = nextItem ? nextItem.tick : midiNotation.endTick;
+			console.assert(item.endTick > item.tick, "invalid tick item, tick span is non-positive:", item, tokenTable[item.tick]);
 
 			if (nextItem && item.row === nextItem.row && [0, 1].includes(nextItem.measure - item.measure))
 				item.endX = nextItem.x;