+++ changelog.md
@@ -1,6 +1,5 @@
 
 ## TODO
-	assign track on SheetDocument.alignTokensWithNotation
 	MIDI post processing
 		arpeggio
 	implicit notes matching with sheet document
@@ -18,6 +17,11 @@
 		system splitting, chord mode symbol affiliation
 		remove empty track	(6.1 落叶.ly)
 
+
+
+*	2020.9.15
+
+	+	lilyInterpreter.ts: fix music track anchro pitch for single chord non-relative music block.
 
 
 *	2020.9.14